+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes:
 
 - Add missing barcode ID to export
 - Don't show update barcode dialog if value is the same as card ID
+- Add Finnish translation
 
 ## v1.14 (2021-06-07)